requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
diffusers==0.29.0
transformers==4.41.2
accelerate==0.30.1
safetensors==0.4.3
Pillow==10.3.0
